Cleaned up the code and fixed the joystick function to use floating point

+++ Math.c
@@ -1,7 +1,7 @@
 #pragma once
 #include "Math.h"
 
-int signOf(int x)
+float signOf(float x)
 {
 	return (x / abs(x));
 }
